cleanup on public proposals component
