+++ src/components/PublicProposalsPage/PublicProposals.js
@@ -1,27 +1,19 @@
 import React, { useState, useEffect } from "react";
+import PropTypes from "prop-types";
 import "./styles.css";
 import { Tabs, Tab } from "../Tabs";
 import LazyList from "../LazyList/LazyList";
 import ThingLinkProposal from "../snew/ThingLink/ThingLinkProposal";
 import {
-  PROPOSAL_VOTING_NOT_AUTHORIZED,
-  PROPOSAL_VOTING_AUTHORIZED,
-  PROPOSAL_VOTING_ACTIVE,
-  PROPOSAL_VOTING_FINISHED,
-  PROPOSAL_STATUS_ABANDONED
-} from "../../constants";
+  tabValues,
+  getProposalTokensByTabOption,
+  getProposalsByTabOption
+} from "./helpers";
 import { proposalToT3 } from "../../lib/snew";
 import proposalsConnector from "../../connectors/publicProposals";
 import { PageLoadingIcon } from "../snew";
 
-const PAGE_SIZE = 6;
-
-const tabValues = {
-  IN_DISCUSSSION: 0,
-  VOTING: 1,
-  FINISHED: 2,
-  ABANDONED: 3
-};
+const DEFAULT_PAGE_SIZE = 6;
 
 const renderProposal = (record, index) => (
   <ThingLinkProposal
@@ -29,40 +21,6 @@ const renderProposal = (record, index) => (
     {...{ ...proposalToT3(record, index).data, comments: [] }}
   />
 );
-
-const getProposalTokensByTabOption = (tabOption, proposalsTokens) => {
-  if (!proposalsTokens) return [];
-  const { pre, active, finished, abandoned } = proposalsTokens;
-  switch (tabOption) {
-    case tabValues.IN_DISCUSSSION:
-      return pre;
-    case tabValues.VOTING:
-      return active;
-    case tabValues.FINISHED:
-      return finished;
-    case tabValues.ABANDONED:
-      return abandoned;
-    default:
-      return [];
-  }
-};
-
-const getProposalsByTabOption = (tabOption, proposals, getVoteStatus) => {
-  const proposalVotingStatus = proposal =>
-    getVoteStatus(proposal.censorshiprecord.token).status;
-  const mapTabOptionToFilter = {
-    [tabValues.IN_DISCUSSSION]: p =>
-      proposalVotingStatus(p) === PROPOSAL_VOTING_NOT_AUTHORIZED ||
-      proposalVotingStatus(p) === PROPOSAL_VOTING_AUTHORIZED,
-    [tabValues.VOTING]: p => proposalVotingStatus(p) === PROPOSAL_VOTING_ACTIVE,
-    [tabValues.FINISHED]: p =>
-      proposalVotingStatus(p) === PROPOSAL_VOTING_FINISHED,
-    [tabValues.ABANDONED]: p => p.status === PROPOSAL_STATUS_ABANDONED
-  };
-  const filter = mapTabOptionToFilter[tabOption];
-
-  return proposals.filter(filter);
-};
 
 const PlaceHolder = () => <div className="card-placeholder" />;
 
@@ -82,7 +40,8 @@ const PublicProposals = ({
   onFetchProposalsVoteStatus,
   onFetchTokenInventory,
   isLoading,
-  error
+  error,
+  pageSize
 }) => {
   const [tabOption, setTabOption] = useState(tabValues.IN_DISCUSSSION);
   const [hasMoreToLoad, setHasMore] = useState(true);
@@ -104,10 +63,7 @@ const PublicProposals = ({
 
   const handleFetchMoreRecords = async () => {
     const index = filteredProposals.length;
-    const propTokensToBeFetched = filteredTokens.slice(
-      index,
-      index + PAGE_SIZE
-    );
+    const propTokensToBeFetched = filteredTokens.slice(index, index + pageSize);
     setHasMore(false);
     setItemsOnLoad(propTokensToBeFetched.length);
     await onFetchVettedByTokens(propTokensToBeFetched);
@@ -166,6 +122,22 @@ const PublicProposals = ({
       )}
     </>
   );
+};
+
+PublicProposals.propTypes = {
+  proposals: PropTypes.array,
+  proposalsTokens: PropTypes.array,
+  getVoteStatus: PropTypes.func,
+  onFetchVettedByTokens: PropTypes.func.isRequired,
+  onFetchProposalsVoteStatus: PropTypes.func.isRequired,
+  onFetchTokenInventory: PropTypes.func.isRequired,
+  isLoading: PropTypes.bool,
+  error: PropTypes.string,
+  pageSize: PropTypes.number
+};
+
+PublicProposals.defaultProps = {
+  pageSize: DEFAULT_PAGE_SIZE
 };
 
 export default proposalsConnector(PublicProposals);
